datalogger: Change flags from CODAL-only to both DAL and CODAL, due to issue #5352.
Due to issue https://github.com/microsoft/pxt-microbit/issues/5352
the flags set up only for CODAL don't work correctly.

The issue will be fixed in the future, but as a workaround we can
set the flag for both DAL and CODAL, even if this is a V2 only
feature.

Also removes the user config switch, as it didn't work and we
prefer having less options.
The flag can always be manually changes in the pxt.json file.