Remove rng seed call in platform_init().
It is not needed, as the random number generator is already seeded
in uBit.init().
This also works around this issue until it is resolved in DAL &
CODAL:
https://github.com/microsoft/pxt-microbit/issues/5409

Apart from that it is a costly call, specially in V1 where it
takes 50 microsec or more, and in theory can take up to 480 us.